main/requirements.txt
aioamqp==0.14.0
alembic==1.4.2
boto3==1.14.14
botocore==1.17.14
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
click==6.6
docutils==0.15.2
Flask==0.11.1
Flask-Migrate==1.8.0
Flask-Script==2.0.5
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.12
google-api-core==1.21.0
google-auth==1.18.0
google-cloud-core==1.3.0
google-cloud-storage==1.29.0
google-resumable-media==0.5.1
googleapis-common-protos==1.52.0
idna==2.10
itsdangerous==0.24
Jinja2==2.8
jmespath==0.10.0
Mako==1.1.3
MarkupSafe==0.23
pamqp==2.3.0
protobuf==3.12.2
psycopg2-binary==2.8.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-dateutil==2.8.1
python-editor==1.0.4
python-slugify==1.2.0
pytz==2020.1
redis==3.5.3
requests==2.24.0
rsa==4.6
s3transfer==0.3.3
six==1.15.0
SQLAlchemy==1.3.17
Unidecode==1.1.1
urllib3==1.25.9
Werkzeug==0.11.10
WTForms==2.1
